feat: extract error fallback to component
